vanna-ai/check_schema.py
import psycopg2
from dotenv import load_dotenv
import os

load_dotenv()

conn = psycopg2.connect(
    host=os.getenv("DB_HOST"),
    port=os.getenv("DB_PORT"),
    database=os.getenv("DB_NAME"),
    user=os.getenv("DB_USER"),
    password=os.getenv("DB_PASSWORD")
)

cursor = conn.cursor()
cursor.execute("""
    SELECT column_name, data_type 
    FROM information_schema.columns 
    WHERE table_name = 'invoices'
    ORDER BY ordinal_position;
""")

print("Invoices table columns:")
for row in cursor.fetchall():
    print(f"  {row[0]} ({row[1]})")

conn.close()
